\name{AggregateBayes-methods}
\docType{methods}
\alias{AggregateBayes-methods}
\alias{AggregateBayes,RepeatRanking,StabilityLm-method}
\alias{AggregateBayes,RepeatRanking,StabilityOverlap-method}
\title{Bayesian aggregation of repeated rankings}
\description{
 Aggregates rankings over perturbed datasets.
}
\section{Methods}{
\describe{

\item{RR = "RepeatRanking", S = "StabilityLm"}{signature 1}

\item{RR = "RepeatRanking", S = "StabilityOverlap"}{signature 2}
}
For further argument and output information, consult \link{AggregateBayes}.

}

\keyword{univar}

