\name{AggregatePCA-methods}
\docType{methods}
\alias{AggregatePCA-methods}
\alias{AggregatePCA,RepeatRanking-method}
\title{aggregation of repeated rankings by principal components}
\description{
 A principal components analysis is applied to the matrix storing
 the different rankings for each gene. The first principal component
 is then used for aggregation.
}
\section{Methods}{
\describe{

\item{RR = "RepeatRanking"}{signature 1}

}
For further argument and output information, consult \link{AggregatePCA}.

}

\keyword{univar}