\name{AggregatePenalty-methods}
\docType{methods}
\alias{AggregatePenalty-methods}
\alias{AggregatePenalty,RepeatRanking-method}
\title{aggregation of repeated rankings by a variance penalty}
\description{
  The idea behind this form of aggregation is to find 'reliable'
  candidate genes, i.e. those ones that are highly ranked and little
  variable at the same time. Higher variability is stronger penalized.
}
\section{Methods}{
\describe{

\item{RR = "RepeatRanking"}{signature 1}

}
For further argument and output information, consult \link{AggregatePenalty}.

}

\keyword{univar}