\name{AggregateSimple-methods}
\docType{methods}
\alias{AggregateSimple-methods}
\alias{AggregateSimple,RepeatRanking,StabilityLm-method}
\alias{AggregateSimple,RepeatRanking,StabilityOverlap-method}
\title{'Simple' aggregation of repeated rankings}
\description{Aggregates rankings over perturbed datasets.}
\section{Methods}{
\describe{

\item{RR = "RepeatRanking", S = "StabilityLm"}{signature 1}

\item{RR = "RepeatRanking", S = "StabilityOverlap"}{signature 2}
}
For further argument and output information, consult \link{AggregateSimple}.
}
\keyword{univar}

