\name{GeneSelector-methods}
\docType{methods}
\alias{GeneSelector-methods}
\alias{GeneSelector,list-method}
\title{Exclude genes from being candidates for differential expression}
\description{
 For a detailed description, s. \link{GeneSelector}.
}
\section{Methods}{
 The input is a list of objects of class \code{GeneSelector} or 
 \code{AggregatedRanking}. 
\describe{
\item{Rlist = "list"}{signature 1}
}

For further argument and output information, consult
 \link{GeneSelector}.
}

\keyword{univar}
