\name{GetStabilityOverlap-methods}
\docType{methods}
\alias{GetStabilityOverlap-methods}
\alias{GetStabilityOverlap,RepeatRanking-method}
\title{Stability measures for gene rankings}
\description{
  The similarity of two ordered genelists is assessed
  by a 'weighted cumulative sum' of the number of 
  overlaps up to a certain position in the list and is called
  'Overlap Score'.   
}
\section{Methods}{
The input is an object of class \code{RepeatRanking}.
\describe{
\item{RR = "RepeatRanking"}{signature 1}
}
For further argument and output information, consult \link{GetStabilityOverlap}.
}
\keyword{univar}

