\name{HeatmapMethods-methods}
\docType{methods}
\alias{HeatmapMethods-methods}
\alias{HeatmapMethods,list-method}
\title{Heatmap of genes and ranking procedures}
\description{
 For a detailed description, s. \link{HeatmapMethods}.
}
\section{Methods}{
The input is a list of objects of class \code{GeneRanking} or 
\code{AggregatedRanking}.
\describe{
\item{Rlist = "list"}{signature 1}
}
For further argument and output information, consult
 \link{HeatmapMethods}.
}

\keyword{univar}

