\name{join-methods}
\docType{methods}
\alias{join-methods}
\alias{join,RepeatRanking,RepeatRanking-method}
\title{Combine two objects of class RepeatRankings}
\description{
 For a detailed description, s. \link{join}.
}
\section{Methods}{         
The inputs are two objects of class \code{RepeatRanking}
\describe{
\item{RR1 = "RepeatRanking", RR2 = "RepeatRanking"}{signature 1}
}}
\keyword{univar}

