\name{PCAMethods-methods}
\docType{methods}
\alias{PCAMethods-methods}
\alias{PCAMethods,list-method}
\title{Principal Components Analysis for different ranking procedures}
\description{
 For a detailed description, s. \link{PCAMethods}.
}
\section{Methods}{
The input is a list of objects of class \code{GeneRanking}.
\describe{
\item{Rlist = "list"}{signature 1}
}
For further argument and output information, consult
 \link{PCAMethods}.
}

\keyword{univar}

