\name{RankingGap-methods}
\docType{methods}
\alias{RankingGap-methods}
\alias{RankingGap,matrix,numeric-method}
\alias{RankingGap,matrix,factor-method}
\alias{RankingGap,ExpressionSet,character-method}
\title{Ranking based on 'gaps'.}
\description{
 For a detailed description, s. \link{RankingGap}.
}
\section{Methods}{
The input (gene expression and class labels) 
can be given in three different ways:
\describe{

\item{x = "matrix", y = "numeric"}{signature 1}

\item{x = "matrix", y = "factor"}{signature 2}

\item{x = "ExpressionSet", y = "character"}{signature 3}
}
For further argument and output information, consult 
\link{RankingGap}.
}
\keyword{univar}

