\name{toplist-methods}
\docType{methods}
\alias{toplist-methods}
\alias{toplist}

\title{'Toplist' methods}
\description{
 Several code \code{toplists} methods are defined, s. below.
}
\section{Methods}{
\describe{

\item{object = "GeneRanking"}{s. \link{GeneRanking-class}}

\item{object = "RepeatRanking"}{s. \link{GeneRanking-class}}

\item{object = "CombinedRanking"}{s. \link{CombinedRanking-class}}
}}

\keyword{univar}


