\name{toplist-methods}
\docType{methods}
\alias{toplist-methods}
\alias{toplist}

\title{'Toplist' methods}
\description{
 Several code \code{toplists} methods are defined, s. below.
}
\section{Methods}{
\describe{

\item{object = "GeneRanking"}{s. \link{GeneRanking-class}}

\item{object = "RepeatedRanking"}{s. \link{RepeatedRanking-class}}

\item{object = "AggregatedRanking"}{s. \link{AggregatedRanking-class}}

\item{object = "GeneSelectorOutput"}{s. \link{GeneSelectorOutput-class}}
}}

\keyword{univar}


